Generate a custom selector function

Usage

.generic_selector(variable_column, select_column, select_expr, fun_name)

.is_selector_scoped(variable_column, select_column)

Arguments

variable_column

string indicating column variable names are stored

select_column

character vector of columns used in the select_expr= argument

select_expr

unquoted predicate command to subset a data frame to select variables

fun_name

quoted name of function where .generic_selector() is being used. This helps with error messaging.

Value

custom selector functions

Details

.is_selector_scoped() checks if a selector has been properly registered in env_variable_type$df_var_info.