model_get_model_matrix {broom.helpers}	R Documentation

Get the model matrix of a model

Description

The structure of the object returned by stats::model.matrix() could slightly differ for certain types of models. model_get_model_matrix() will always return an object with the same structure as stats::model.matrix.default().

Usage

model_get_model_matrix(model, ...)

## Default S3 method:
model_get_model_matrix(model, ...)

## S3 method for class 'multinom'
model_get_model_matrix(model, ...)

## S3 method for class 'clm'
model_get_model_matrix(model, ...)

## S3 method for class 'brmsfit'
model_get_model_matrix(model, ...)

## S3 method for class 'glmmTMB'
model_get_model_matrix(model, ...)

Arguments

model
a model object

...
additional arguments passed to stats::model.matrix()

Details

For models fitted with glmmTMB::glmmTMB(), it will return a model matrix taking into account all components ("cond", "zi" and "disp"). For a more restricted model matrix, please refer to glmmTMB::model.matrix.glmmTMB().

See Also

stats::model.matrix()

Other model_helpers: model_compute_terms_contributions(), model_get_assign(), model_get_coefficients_type(), model_get_contrasts(), model_get_model_frame(), model_get_model(), model_get_nlevels(), model_get_n(), model_get_offset(), model_get_response(), model_get_terms(), model_get_weights(), model_get_xlevels(), model_identify_variables(), model_list_contrasts(), model_list_terms_levels(), model_list_variables()

Examples

lm(hp ~ mpg + factor(cyl), mtcars) %>%
  model_get_model_matrix() %>%
  head()

[Package broom.helpers version 1.3.0.9001 Index]