Get the number of observations

model_get_n(model)

# S3 method for default
model_get_n(model)

# S3 method for glm
model_get_n(model)

# S3 method for gam
model_get_n(model)

# S3 method for glmerMod
model_get_n(model)

# S3 method for multinom
model_get_n(model)

# S3 method for coxph
model_get_n(model)

# S3 method for survreg
model_get_n(model)

Arguments

model
a model object

Details

For binomial and multinomial logistic models, will also return the number of events.

For Poisson models, will return the number of events and exposure time (defined with stats::offset()).

For Cox models (survival::coxph()), will return the number of events and exposure time.

See tidy_add_n() for more details.

The total number of observations (N_obs), of events (N_event) and of exposure time (Exposure) are stored as attributes of the returned tibble.

This function does not cover lavaan models (NULL is returned).

See also

Examples

lm(hp ~ mpg + factor(cyl) + disp:hp, mtcars) %>%
  model_get_n()

#> # A tibble: 6 x 2
#>   term         n_obs
#>   <chr>        <dbl>
#> 1 (Intercept)     32
#> 2 mpg             32
#> 3 factor(cyl)6     7
#> 4 factor(cyl)8    14
#> 5 hp:disp         32
#> 6 factor(cyl)4    11

mod <- glm(
  response ~ stage * grade + trt,
  gtsummary::trial,
  family = binomial,
  contrasts = list(stage = contr.sum, grade = contr.treatment(3, 2), trt = "contr.SAS")
)
mod %>% model_get_n()

#> # A tibble: 16 x 3
#>    term          n_obs n_event
#>    <chr>         <dbl>   <dbl>
#>  1 (Intercept)     193      61
#>  2 stage1           52      18
#>  3 stage2           52      13
#>  4 stage3           40      15
#>  5 grade1           67      21
#>  6 grade3           63      21
#>  7 trtDrug A        95      28
#>  8 stage1:grade1    17       6
#>  9 stage2:grade1    17       4
#> 10 stage3:grade1    18       7
#> 11 stage1:grade3    13       6
#> 12 stage2:grade3    19       5
#> 13 stage3:grade3    13       4
#> 14 stage4           49      15
#> 15 grade2           63      19
#> 16 trtDrug B        98      33

mod <- glm(
  Survived ~ Class * Age + Sex, data = Titanic %>% as.data.frame(),
  weights = Freq, family = binomial
)
mod %>% model_get_n()

#> # A tibble: 12 x 3
#>    term               n_obs n_event
#>    <chr>              <dbl>   <dbl>
#>  1 (Intercept)         2201     711
#>  2 Class2nd             285     118
#>  3 Class3rd             706     178
#>  4 ClassCrew            885     212
#>  5 AgeAdult            2092     654
#>  6 SexFemale            470     344
#>  7 Class2nd:AgeAdult    261      94
#>  8 Class3rd:AgeAdult    627     151
#>  9 ClassCrew:AgeAdult   885     212
#> 10 Class1st             325     203
#> 11 AgeChild             109      57
#> 12 SexMale             1731     367

d <- dplyr::as_tibble(Titanic) %>%
  dplyr::group_by(Class, Sex, Age) %>%
  dplyr::summarise(
    n_survived = sum(n * (Survived == "Yes")),
    n_dead = sum(n * (Survived == "No"))
  )

#> `summarise()` regrouping output by 'Class', 'Sex' (override with `.groups` argument)

mod <- glm(cbind(n_survived, n_dead) ~ Class * Age + Sex, data = d, family = binomial)
mod %>% model_get_n()

#> # A tibble: 12 x 3
#>    term               n_obs n_event
#>    <chr>              <dbl>   <dbl>
#>  1 (Intercept)         2201     711
#>  2 Class2nd             285     118
#>  3 Class3rd             706     178
#>  4 ClassCrew            885     212
#>  5 AgeChild             109      57
#>  6 SexMale             1731     367
#>  7 Class2nd:AgeChild     24      24
#>  8 Class3rd:AgeChild     79      27
#>  9 ClassCrew:AgeChild     0       0
#> 10 Class1st             325     203
#> 11 AgeAdult            2092     654
#> 12 SexFemale            470     344

mod <- glm(response ~ age + grade * trt, gtsummary::trial, family = poisson)
mod %>% model_get_n()

#> # A tibble: 9 x 4
#>   term               n_obs n_event exposure
#>   <chr>              <dbl>   <dbl>    <dbl>
#> 1 (Intercept)          183      58      183
#> 2 age                  183      58      183
#> 3 gradeII               58      17       58
#> 4 gradeIII              60      20       60
#> 5 trtDrug B             94      31       94
#> 6 gradeII:trtDrug B     29      10       29
#> 7 gradeIII:trtDrug B    33       8       33
#> 8 gradeI                65      21       65
#> 9 trtDrug A             89      27       89

mod <- glm(
  response ~ trt * grade + offset(ttdeath),
  gtsummary::trial,
  family = poisson
)
mod %>% model_get_n()

#> # A tibble: 8 x 4
#>   term               n_obs n_event exposure
#>   <chr>              <dbl>   <dbl>    <dbl>
#> 1 (Intercept)          193      61  2.48e12
#> 2 trtDrug B             98      33  1.15e12
#> 3 gradeII               63      19  8.24e11
#> 4 gradeIII              63      21  6.31e11
#> 5 trtDrug B:gradeII     33      12  3.71e11
#> 6 trtDrug B:gradeIII    33       8  2.94e11
#> 7 trtDrug A             95      28  1.33e12
#> 8 gradeI                67      21  1.03e12

df <- survival::lung %>% dplyr::mutate(sex = factor(sex))
mod <- survival::coxph(survival::Surv(time, status) ~ ph.ecog + age + sex, data = df)
mod %>% model_get_n()

#> # A tibble: 5 x 4
#>   term        n_obs n_event exposure
#>   <chr>       <dbl>   <dbl>    <dbl>
#> 1 (Intercept)   227     164    69522
#> 2 ph.ecog       227     164    69522
#> 3 age           227     164    69522
#> 4 sex2           90      53    30507
#> 5 sex1          137     111    39015

mod <- lme4::lmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
mod %>% model_get_n()

#> # A tibble: 2 x 2
#>   term        n_obs
#>   <chr>       <dbl>
#> 1 (Intercept)   180
#> 2 Days          180

mod <- lme4::glmer(response ~ trt * grade + (1 | stage),
  family = binomial, data = gtsummary::trial
)

#> boundary (singular) fit: see ?isSingular

mod %>% model_get_n()

#> # A tibble: 8 x 3
#>   term               n_obs n_event
#>   <chr>              <dbl>   <dbl>
#> 1 (Intercept)          193      61
#> 2 trtDrug B             98      33
#> 3 gradeII               63      19
#> 4 gradeIII              63      21
#> 5 trtDrug B:gradeII     33      12
#> 6 trtDrug B:gradeIII    33       8
#> 7 trtDrug A             95      28
#> 8 gradeI                67      21

mod <- lme4::glmer(cbind(incidence, size - incidence) ~ period + (1 | herd),
  family = binomial, data = lme4::cbpp
)
mod %>% model_get_n()

#> # A tibble: 5 x 3
#>   term        n_obs n_event
#>   <chr>       <dbl>   <dbl>
#> 1 (Intercept)   842      99
#> 2 period2       212      17
#> 3 period3       197      14
#> 4 period4       155       7
#> 5 period1       278      61