model_get_terms {broom.helpers}	R Documentation

Get the terms of a model

Description

Return the result of stats::terms() applied to the model or NULL if it is not possible to get terms from model.

Usage

model_get_terms(model)

## Default S3 method:
model_get_terms(model)

## S3 method for class 'brmsfit'
model_get_terms(model)

## S3 method for class 'glmmTMB'
model_get_terms(model)

Arguments

model
a model object

Details

For models fitted with glmmTMB::glmmTMB(), it will return a terms object taking into account all components ("cond" and "zi"). For a more restricted terms object, please refer to glmmTMB::terms.glmmTMB().

See Also

stats::terms()

Other model_helpers: model_compute_terms_contributions(), model_get_assign(), model_get_coefficients_type(), model_get_contrasts(), model_get_model_frame(), model_get_model_matrix(), model_get_model(), model_get_nlevels(), model_get_n(), model_get_offset(), model_get_response(), model_get_weights(), model_get_xlevels(), model_identify_variables(), model_list_contrasts(), model_list_terms_levels(), model_list_variables()

Examples

lm(hp ~ mpg + factor(cyl), mtcars) %>%
  model_get_terms()

[Package broom.helpers version 1.3.0.9001 Index]