model_list_contrasts {broom.helpers}	R Documentation

List contrasts used by a model

Description

List contrasts used by a model

Usage

model_list_contrasts(model)

## Default S3 method:
model_list_contrasts(model)

Arguments

model
a model object

Details

For models with no intercept, no contrasts will be applied to one of the categorical variable. In such case, one dummy term will be returned for each level of the categorical variable.

Value

A tibble with three columns:

variable: variable name

contrasts: contrasts used

contrasts_type: type of contrasts ("treatment", "sum", "poly", "helmert", "other" or "no.contrast")

reference: for variables with treatment, SAS or sum contrasts, position of the reference level

See Also

Other model_helpers: model_compute_terms_contributions(), model_get_assign(), model_get_coefficients_type(), model_get_contrasts(), model_get_model_frame(), model_get_model_matrix(), model_get_model(), model_get_nlevels(), model_get_n(), model_get_offset(), model_get_response(), model_get_terms(), model_get_weights(), model_get_xlevels(), model_identify_variables(), model_list_terms_levels(), model_list_variables()

Examples

glm(
  am ~ mpg + factor(cyl),
  data = mtcars,
  family = binomial,
  contrasts = list(`factor(cyl)` = contr.sum)
) %>%
  model_list_contrasts()

[Package broom.helpers version 1.3.0.9001 Index]