model_list_variables {broom.helpers}	R Documentation

List all the variables used in a model

Description

Including variables used only in an interaction.

Usage

model_list_variables(model, labels = NULL, only_variable = FALSE)

## Default S3 method:
model_list_variables(model, labels = NULL, only_variable = FALSE)

## S3 method for class 'lavaan'
model_list_variables(model, labels = NULL, only_variable = FALSE)

Arguments

model
a model object

labels
an optional named list or named vector of custom variable labels

only_variable
if TRUE, will return only "variable" column

Value

A tibble with three columns:

variable: the corresponding variable

var_class: class of the variable (cf. stats::.MFclass())

label_attr: variable label defined in the original data frame with the label attribute (cf. labelled::var_label())

var_label: a variable label (by priority, labels if defined, label_attr if available, otherwise variable)

See Also

Other model_helpers: model_compute_terms_contributions(), model_get_assign(), model_get_coefficients_type(), model_get_contrasts(), model_get_model_frame(), model_get_model_matrix(), model_get_model(), model_get_nlevels(), model_get_n(), model_get_offset(), model_get_response(), model_get_terms(), model_get_weights(), model_get_xlevels(), model_identify_variables(), model_list_contrasts(), model_list_terms_levels()

Examples

Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes"))) %>%
  glm(
    Survived ~ Class + Age : Sex,
    data = ., weights = .$n,
    family = binomial
  ) %>%
  model_list_variables()

iris %>%
  lm(
    Sepal.Length ~ poly(Sepal.Width, 2) + Species,
    data = .,
    contrasts = list(Species = contr.sum)
  ) %>%
  model_list_variables()

if (requireNamespace("gtsummary")) {
  glm(
    response ~ poly(age, 3) + stage + grade * trt,
    na.omit(gtsummary::trial),
    family = binomial,
  ) %>%
    model_list_variables()
}

[Package broom.helpers version 1.3.0.9001 Index]