Listing of Supported Models

Listing of Supported Models

Usage

supported_models

Format

A data frame with one row per supported model

model

Model

notes

Notes

Supported models

model	notes

betareg::betareg()	Use tidy_parameters() as tidy_fun with component argument to control with coefficients to return. broom::tidy() does not support the exponentiate argument for betareg models, use tidy_parameters() instead.

biglm::bigglm()

brms::brm()	broom.mixed package required

cmprsk::crr()	Limited support. It is recommended to use tidycmprsk::crr() instead.

fixest::feglm()	May fail with R <= 4.0.

fixest::femlm()	May fail with R <= 4.0.

fixest::feNmlm()	May fail with R <= 4.0.

fixest::feols()	May fail with R <= 4.0.

gam::gam()

geepack::geeglm()

glmmTMB::glmmTMB()	broom.mixed package required

glmtoolbox::glmgee()

lavaan::lavaan()	Limited support for categorical variables

lfe::felm()

lme4::glmer.nb()	broom.mixed package required

lme4::glmer()	broom.mixed package required

lme4::lmer()	broom.mixed package required

logitr::logitr()	Requires logitr >= 0.8.0

MASS::glm.nb()

MASS::polr()

mgcv::gam()	Use default tidier broom::tidy() for smooth terms only, or gtsummary::tidy_gam() to include parametric terms

mice::mira	Limited support. If mod is a mira object, use tidy_fun = function(x, ...) {mice::pool(x) &#124;> mice::tidy(...)}

mmrm::mmrm()

multgee::nomLORgee()	Use tidy_multgee() as tidy_fun.

multgee::ordLORgee()	Use tidy_multgee() as tidy_fun.

nnet::multinom()

ordinal::clm()	Limited support for models with nominal predictors.

ordinal::clmm()	Limited support for models with nominal predictors.

parsnip::model_fit	Supported as long as the type of model and the engine is supported.

plm::plm()

pscl::hurdle()	Use tidy_zeroinfl() as tidy_fun.

pscl::zeroinfl()	Use tidy_zeroinfl() as tidy_fun.

rstanarm::stan_glm()	broom.mixed package required

stats::aov()	Reference rows are not relevant for such models.

stats::glm()

stats::lm()

stats::nls()	Limited support

survey::svycoxph()

survey::svyglm()

survey::svyolr()

survival::cch()	Experimental support.

survival::clogit()

survival::coxph()

survival::survreg()

tidycmprsk::crr()

VGAM::vgam()	Experimental support. It is recommended to use tidy_vgam() as tidy_fun.

VGAM::vglm()	Experimental support. It is recommended to use tidy_vgam() as tidy_fun.