Add an estimate value to references rows for categorical variables

For categorical variables with a treatment contrast (stats::contr.treatment()) or a SAS contrast (stats::contr.SAS()) will add an estimate equal to 0 (or 1 if exponentiate = TRUE) to the reference row.

tidy_add_estimate_to_reference_rows(
  x,
  exponentiate = FALSE,
  model = tidy_get_model(x)
)

Arguments

x
a tidy tibble

exponentiate
logical indicating whether or not to exponentiate the coefficient estimates. It should be consistent with the original call to broom::tidy()

model
the corresponding model, if not attached to x

Details

For categorical variables with a sum contrast (stats::contr.sum()), the estimate value of the reference row will be equal to the sum of all other coefficients multiplied by -1 (eventually exponentiated if exponentiate = TRUE), and obtained with stats::dummy.coef(). To be noted, for sum contrasts, the model coefficients correspond to the difference of each level with the great mean.

For other variables, no change will be made.

If the reference_row column is not yet available in x, tidy_add_reference_rows() will be automatically applied.

See also

Examples

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(dplyr::across(where(is.character), factor))

df %>%
  glm(
    Survived ~ Class + Age + Sex,
    data = ., weights = .$n, family = binomial,
    contrasts = list(Age = contr.sum, Class = "contr.SAS")
  ) %>%
  tidy_and_attach(exponentiate = TRUE) %>%
  tidy_add_reference_rows() %>%
  tidy_add_estimate_to_reference_rows(exponentiate = TRUE)

#> # A tibble: 9 x 10
#>   term  variable var_class var_type estimate std.error statistic   p.value
#>   <chr> <chr>    <chr>     <chr>       <dbl>     <dbl>     <dbl>     <dbl>
#> 1 (Int… NA       NA        interce…  NA          0.206     8.34   7.14e-17
#> 2 Clas… Class    factor    categor…   2.36       0.157     5.45   5.00e- 8
#> 3 Clas… Class    factor    categor…   0.852      0.174    -0.923  3.56e- 1
#> 4 Clas… Class    factor    categor…   0.398      0.149    -6.19   5.93e-10
#> 5 Clas… Class    factor    categor…   1         NA        NA     NA       
#> 6 Age1  Age      factor    categor…   0.588      0.122    -4.35   1.36e- 5
#> 7 Age2  Age      factor    categor…   1.70      NA        NA     NA       
#> 8 SexF… Sex      factor    categor…   1         NA        NA     NA       
#> 9 SexM… Sex      factor    categor…   0.0889     0.140   -17.2    1.43e-66
#> # … with 2 more variables: contrasts <chr>, reference_row <lgl>

if(requireNamespace("gtsummary")) {
  glm(
    response ~ stage + grade * trt,
    gtsummary::trial,
    family = binomial,
    contrasts = list(stage = contr.treatment(4, base = 3),
                     grade = contr.treatment(3, base = 2),
                     trt = contr.treatment(2, base = 2))
  ) %>%
    tidy_and_attach() %>%
    tidy_add_reference_rows() %>%
    tidy_add_estimate_to_reference_rows
}

#> Loading required namespace: gtsummary

#> # A tibble: 12 x 10
#>    term  variable var_class var_type estimate std.error statistic p.value
#>    <chr> <chr>    <chr>     <chr>       <dbl>     <dbl>     <dbl>   <dbl>
#>  1 (Int… NA       NA        interce…   NA         0.487    -0.402  0.688 
#>  2 stag… stage    factor    categor…   -0.138     0.455    -0.303  0.762 
#>  3 stag… stage    factor    categor…   -0.698     0.471    -1.48   0.138 
#>  4 stag… stage    factor    categor…    0        NA        NA     NA     
#>  5 stag… stage    factor    categor…   -0.447     0.466    -0.959  0.338 
#>  6 grad… grade    factor    categor…    0.142     0.516     0.274  0.784 
#>  7 grad… grade    factor    categor…    0        NA        NA     NA     
#>  8 grad… grade    factor    categor…   -0.615     0.550    -1.12   0.264 
#>  9 trt1  trt      character categor…   -0.710     0.575    -1.23   0.217 
#> 10 trt2  trt      character categor…    0        NA        NA     NA     
#> 11 grad… grade:t… NA        interac…   NA         0.794    -0.227  0.821 
#> 12 grad… grade:t… NA        interac…   NA         0.803     2.04   0.0418
#> # … with 2 more variables: contrasts <chr>, reference_row <lgl>