Add header rows variables with several terms

For variables with several terms (usually categorical variables but could also be the case of continuous variables with polynomial terms or splines), tidy_add_header_rows() will add an additional row per variable, where label will be equal to var_label. These additional rows could be identified with header_row column.

tidy_add_header_rows(x, show_single_row = NULL, model = tidy_get_model(x))

Arguments

x
a tidy tibble

show_single_row
a vector indicating the names of binary variables that should be displayed on a single row

model
the corresponding model, if not attached to x

Details

In addition, the show_single_row argument allows to specify a list of dichotomous variables that should be displayed on a single row instead of two rows. This argument will have no effect if reference rows have not been added to the tibble (cf. tidy_add_reference_rows()).

If the label column is not yet available in x, tidy_add_term_labels() will be automatically applied.

See also

Examples

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))

df %>%
  glm(
    Survived ~ Class + Age + Sex,
    data = ., weights = .$n, family = binomial,
    contrasts = list(Age = contr.sum, Class = "contr.SAS")
  ) %>%
  tidy_and_attach() %>%
  tidy_add_variable_labels(labels = list(Class = "Custom label for Class")) %>%
  tidy_add_reference_rows() %>%
  tidy_add_header_rows()

#> # A tibble: 12 x 13
#>    term  variable var_label var_class var_type header_row contrasts
#>    <chr> <chr>    <chr>     <chr>     <chr>    <lgl>      <chr>    
#>  1 (Int… NA       (Interce… NA        interce… NA         NA       
#>  2 NA    Class    Custom l… character categor… TRUE       NA       
#>  3 Clas… Class    Custom l… character categor… FALSE      contr.SAS
#>  4 Clas… Class    Custom l… character categor… FALSE      contr.SAS
#>  5 Clas… Class    Custom l… character categor… FALSE      contr.SAS
#>  6 Clas… Class    Custom l… character categor… FALSE      contr.SAS
#>  7 NA    Age      Age       character categor… TRUE       NA       
#>  8 Age1  Age      Age       character categor… FALSE      contr.sum
#>  9 Age2  Age      Age       character categor… FALSE      contr.sum
#> 10 NA    Sex      Sex       character categor… TRUE       NA       
#> 11 SexF… Sex      Sex       character categor… FALSE      contr.tr…
#> 12 SexM… Sex      Sex       character categor… FALSE      contr.tr…
#> # … with 6 more variables: reference_row <lgl>, label <chr>, estimate <dbl>,
#> #   std.error <dbl>, statistic <dbl>, p.value <dbl>

if (requireNamespace("gtsummary")) {
  glm(
    response ~ stage + grade * trt,
    gtsummary::trial,
    family = binomial,
    contrasts = list(
      stage = contr.treatment(4, base = 3),
      grade = contr.treatment(3, base = 2),
      trt = contr.treatment(2, base = 2)
    )
  ) %>%
    tidy_and_attach() %>%
    tidy_add_reference_rows() %>%
    tidy_add_header_rows()
}

#> # A tibble: 16 x 13
#>    term  variable var_label var_class var_type header_row contrasts
#>    <chr> <chr>    <chr>     <chr>     <chr>    <lgl>      <chr>    
#>  1 (Int… NA       (Interce… NA        interce… NA         NA       
#>  2 NA    stage    T Stage   factor    categor… TRUE       NA       
#>  3 stag… stage    T Stage   factor    categor… FALSE      contr.tr…
#>  4 stag… stage    T Stage   factor    categor… FALSE      contr.tr…
#>  5 stag… stage    T Stage   factor    categor… FALSE      contr.tr…
#>  6 stag… stage    T Stage   factor    categor… FALSE      contr.tr…
#>  7 NA    grade    Grade     factor    categor… TRUE       NA       
#>  8 grad… grade    Grade     factor    categor… FALSE      contr.tr…
#>  9 grad… grade    Grade     factor    categor… FALSE      contr.tr…
#> 10 grad… grade    Grade     factor    categor… FALSE      contr.tr…
#> 11 NA    trt      Chemothe… character categor… TRUE       NA       
#> 12 trt1  trt      Chemothe… character categor… FALSE      contr.SAS
#> 13 trt2  trt      Chemothe… character categor… FALSE      contr.SAS
#> 14 NA    grade:t… Grade * … NA        interac… TRUE       NA       
#> 15 grad… grade:t… Grade * … NA        interac… FALSE      NA       
#> 16 grad… grade:t… Grade * … NA        interac… FALSE      NA       
#> # … with 6 more variables: reference_row <lgl>, label <chr>, estimate <dbl>,
#> #   std.error <dbl>, statistic <dbl>, p.value <dbl>