Add the (weighted) number of observations

Add the number of observations in a new column n_obs, taking into account any weights if they have been defined.

tidy_add_n(x, model = tidy_get_model(x))

Arguments

x
a tidy tibble

model
the corresponding model, if not attached to x

Details

For continuous variables, it corresponds to all valid observations contributing to the model.

For categorical variables coded with treatment or sum contrasts, each model term could be associated to only one level of the original categorical variable. Therefore, n_obs will correspond to the number of observations associated with that level. n_obs will also be computed for reference rows. For polynomial contrasts (defined with stats::contr.poly()), all levels will contribute to the computation of each model term. Therefore, n_obs will be equal to the total number of observations. For Helmert and custom contrasts, only rows contributing positively (i.e. with a positive contrast) to the computation of a term will be considered for estimating n_obs. The result could therefore be difficult to interpret. For a better understanding of which observations are taken into account to compute n_obs values, you could look at model_compute_terms_contributions().

For interaction terms, only rows contributing to all the terms of the interaction will be considered to compute n_obs.

For binomial logistic models, tidy_add_n() will also return the corresponding number of events (n_event) for each term, taking into account any defined weights. Observed proportions could be obtained as n_obs / n_event.

Similarly, a number of events will be computed for multinomial logistic models (nnet::multinom()) for each level of the outcome (y.level), corresponding to the number of observations equal to that outcome level.

For Poisson models, n_event will be equal to the number of counts per term. In addition, a third column exposure will be computed. If no offset is defined, exposure is assumed to be equal to 1 (eventually multiplied by weights) per observation. If an offset is defined, exposure will be equal to the (weighted) sum of the exponential of the offset (as a reminder, to model the effect of x on the ratio y / z, a Poisson model will be defined as glm(y ~ x + offset(log(z)), family = poisson)). Observed rates could be obtained with n_event / exposure.

For Cox models (survival::coxph()), an individual could be coded with several observations (several rows). n_obs will correspond to the weighted number of observations which could be different from the number of individuals. tidy_add_n() will also compute a (weighted) number of events (n_events) according to the definition of the survival::Surv() object. Exposure time is also returned in exposure column. It is equal to the (weighted) sum of the time variable if only one variable time is passed to survival::Surv(), and to the (weighted) sum of time2 - time if two time variables are defined in survival::Surv().

The (weighted) total number of observations (N_obs), of events (N_event) and of exposure time (Exposure) are stored as attributes of the returned tibble.

See also

Examples

lm(Petal.Length ~ ., data = iris) %>%
  tidy_and_attach() %>%
  tidy_add_n()

#> # A tibble: 6 x 6
#>   term              n_obs estimate std.error statistic  p.value
#>   <chr>             <dbl>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)         150   -1.11     0.270      -4.12 6.45e- 5
#> 2 Sepal.Length        150    0.608    0.0502     12.1  1.07e-23
#> 3 Sepal.Width         150   -0.181    0.0804     -2.25 2.62e- 2
#> 4 Petal.Width         150    0.602    0.121       4.96 1.97e- 6
#> 5 Speciesversicolor    50    1.46     0.173       8.44 3.14e-14
#> 6 Speciesvirginica     50    1.97     0.245       8.06 2.60e-13

lm(Petal.Length ~ ., data = iris, contrasts = list(Species = contr.sum)) %>%
  tidy_and_attach() %>%
  tidy_add_n()

#> # A tibble: 6 x 6
#>   term         n_obs estimate std.error statistic  p.value
#>   <chr>        <dbl>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)    150   0.0349    0.259      0.135 8.93e- 1
#> 2 Sepal.Length   150   0.608     0.0502    12.1   1.07e-23
#> 3 Sepal.Width    150  -0.181     0.0804    -2.25  2.62e- 2
#> 4 Petal.Width    150   0.602     0.121      4.96  1.97e- 6
#> 5 Species1        50  -1.15      0.138     -8.31  6.37e-14
#> 6 Species2        50   0.318     0.0451     7.04  7.19e-11

lm(Petal.Length ~ ., data = iris, contrasts = list(Species = contr.poly)) %>%
  tidy_and_attach() %>%
  tidy_add_n()

#> # A tibble: 6 x 6
#>   term         n_obs estimate std.error statistic  p.value
#>   <chr>        <dbl>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)    150   0.0349    0.259      0.135 8.93e- 1
#> 2 Sepal.Length   150   0.608     0.0502    12.1   1.07e-23
#> 3 Sepal.Width    150  -0.181     0.0804    -2.25  2.62e- 2
#> 4 Petal.Width    150   0.602     0.121      4.96  1.97e- 6
#> 5 Species.L      150   1.40      0.173      8.06  2.60e-13
#> 6 Species.Q      150  -0.389     0.0552    -7.04  7.19e-11

lm(Petal.Length ~ poly(Sepal.Length, 2), data = iris) %>%
  tidy_and_attach() %>%
  tidy_add_n()

#> # A tibble: 3 x 6
#>   term                   n_obs estimate std.error statistic   p.value
#>   <chr>                  <dbl>    <dbl>     <dbl>     <dbl>     <dbl>
#> 1 (Intercept)              150     3.76    0.0685     54.9  7.97e-100
#> 2 poly(Sepal.Length, 2)1   150    18.8     0.839      22.4  3.01e- 49
#> 3 poly(Sepal.Length, 2)2   150    -2.85    0.839      -3.39 8.87e-  4

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))

df %>%
  glm(
    Survived ~ Class + Age + Sex,
    data = ., weights = .$n, family = binomial,
    contrasts = list(Age = contr.sum, Class = "contr.helmert")
  ) %>%
  tidy_and_attach() %>%
  tidy_add_n()

#> # A tibble: 6 x 7
#>   term        n_obs n_event estimate std.error statistic  p.value
#>   <chr>       <dbl>   <dbl>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)  2201     711   1.66      0.162     10.2   1.52e-24
#> 2 Class1        285     118  -0.509     0.0980    -5.19  2.05e- 7
#> 3 Class2        706     178  -0.423     0.0479    -8.83  1.09e-18
#> 4 Class3        885     212   0.0186    0.0303     0.613 5.40e- 1
#> 5 Age1         2092     654  -0.531     0.122     -4.35  1.36e- 5
#> 6 SexMale      1731     367  -2.42      0.140    -17.2   1.43e-66

df %>%
  glm(
    Survived ~ Class * (Age : Sex),
    data = ., weights = .$n, family = binomial,
    contrasts = list(Age = contr.sum, Class = "contr.helmert")
  ) %>%
  tidy_and_attach() %>%
  tidy_add_n()

#> # A tibble: 20 x 7
#>    term                      n_obs n_event  estimate std.error statistic p.value
#>    <chr>                     <dbl>   <dbl>     <dbl>     <dbl>     <dbl>   <dbl>
#>  1 (Intercept)                2201     711  10.3      294.       3.52e-2  0.972 
#>  2 Class1                      285     118   0.333    440.       7.57e-4  0.999 
#>  3 Class2                      706     178  -5.62     147.      -3.83e-2  0.969 
#>  4 Class3                      885     212   0.0848     0.0354   2.40e+0  0.0166
#>  5 AgeAdult:SexFemale          425     316  -8.56     294.      -2.92e-2  0.977 
#>  6 AgeChild:SexFemale           45      28  -0.00790  461.      -1.71e-5  1.00  
#>  7 AgeAdult:SexMale           1667     338 -11.8      294.      -4.03e-2  0.968 
#>  8 AgeChild:SexMale             64      29  NA         NA       NA       NA     
#>  9 Class1:AgeAdult:SexFemale    93      80  -1.20     440.      -2.73e-3  0.998 
#> 10 Class2:AgeAdult:SexFemale   165      76   4.67     147.       3.18e-2  0.975 
#> 11 Class3:AgeAdult:SexFemale    23      20  -0.0450     0.167   -2.70e-1  0.787 
#> 12 Class1:AgeChild:SexFemale    13      13   0.560    692.       8.10e-4  0.999 
#> 13 Class2:AgeChild:SexFemale    31      14   0.402    231.       1.74e-3  0.999 
#> 14 Class3:AgeChild:SexFemale     0       0  NA         NA       NA       NA     
#> 15 Class1:AgeAdult:SexMale     168      14  -1.17     440.      -2.65e-3  0.998 
#> 16 Class2:AgeAdult:SexMale     462      75   5.59     147.       3.81e-2  0.970 
#> 17 Class3:AgeAdult:SexMale     862     192  NA         NA       NA       NA     
#> 18 Class1:AgeChild:SexMale      11      11  NA         NA       NA       NA     
#> 19 Class2:AgeChild:SexMale      48      13  NA         NA       NA       NA     
#> 20 Class3:AgeChild:SexMale       0       0  NA         NA       NA       NA     

glm(response ~ age + grade * trt, gtsummary::trial, family = poisson) %>%
  tidy_and_attach() %>%
  tidy_add_n()

#> # A tibble: 7 x 8
#>   term               n_obs n_event exposure estimate std.error statistic p.value
#>   <chr>              <dbl>   <dbl>    <dbl>    <dbl>     <dbl>     <dbl>   <dbl>
#> 1 (Intercept)          183      58      183 -1.95      0.581     -3.36   7.89e-4
#> 2 age                  183      58      183  0.0113    0.00947    1.19   2.34e-1
#> 3 gradeII               58      17       58  0.00792   0.518      0.0153 9.88e-1
#> 4 gradeIII              60      20       60  0.553     0.458      1.21   2.27e-1
#> 5 trtDrug B             94      31       94  0.511     0.449      1.14   2.55e-1
#> 6 gradeII:trtDrug B     29      10       29 -0.201     0.668     -0.301  7.64e-1
#> 7 gradeIII:trtDrug B    33       8       33 -1.06      0.642     -1.65   9.85e-2

glm(
  response ~ trt * grade + offset(log(ttdeath)),
  gtsummary::trial,
  family = poisson
) %>%
  tidy_and_attach() %>%
  tidy_add_n()

#> # A tibble: 6 x 8
#>   term              n_obs n_event exposure estimate std.error statistic  p.value
#>   <chr>             <dbl>   <dbl>    <dbl>    <dbl>     <dbl>     <dbl>    <dbl>
#> 1 (Intercept)         193      61    3795.  -4.52       0.354  -12.8    2.01e-37
#> 2 trtDrug B            98      33    1853.   0.576      0.449    1.28   2.00e- 1
#> 3 gradeII              63      19    1207.   0.0184     0.518    0.0355 9.72e- 1
#> 4 gradeIII             63      21    1183.   0.728      0.449    1.62   1.05e- 1
#> 5 trtDrug B:gradeII    33      12     576.   0.0548     0.654    0.0838 9.33e- 1
#> 6 trtDrug B:gradeI…    33       8     607.  -1.11       0.635   -1.75   8.00e- 2