Add references rows for categorical variables

For categorical variables with a treatment contrast (stats::contr.treatment()), a SAS contrast (stats::contr.SAS()) or a sum contrast (stats::contr.sum()), add a reference row.

tidy_add_reference_rows(x, model = tidy_get_model(x))

Arguments

x
a tidy tibble

model
the corresponding model, if not attached to x

Details

If the contrasts column is not yet available in x, tidy_add_contrasts() will be automatically applied.

tidy_add_reference_rows() will not populate the label of the reference term. It is therefore better to apply tidy_add_term_labels() after tidy_add_reference_rows() rather than before.

See also

Examples

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))

df %>%
  glm(
    Survived ~ Class + Age + Sex,
    data = ., weights = .$n, family = binomial,
    contrasts = list(Age = contr.sum, Class = "contr.SAS")
  ) %>%
  tidy_and_attach() %>%
  tidy_add_reference_rows()

#> # A tibble: 9 x 10
#>   term  variable var_class var_type contrasts reference_row estimate std.error
#>   <chr> <chr>    <chr>     <chr>    <chr>     <lgl>            <dbl>     <dbl>
#> 1 (Int… NA       NA        interce… NA        NA               1.72      0.206
#> 2 Clas… Class    character categor… contr.SAS FALSE            0.858     0.157
#> 3 Clas… Class    character categor… contr.SAS FALSE           -0.160     0.174
#> 4 Clas… Class    character categor… contr.SAS FALSE           -0.920     0.149
#> 5 Clas… Class    character categor… contr.SAS TRUE            NA        NA    
#> 6 Age1  Age      character categor… contr.sum FALSE           -0.531     0.122
#> 7 Age2  Age      character categor… contr.sum TRUE            NA        NA    
#> 8 SexF… Sex      character categor… contr.tr… TRUE            NA        NA    
#> 9 SexM… Sex      character categor… contr.tr… FALSE           -2.42      0.140
#> # … with 2 more variables: statistic <dbl>, p.value <dbl>

if(requireNamespace("gtsummary")) {
  glm(
    response ~ stage + grade * trt,
    gtsummary::trial,
    family = binomial,
    contrasts = list(stage = contr.treatment(4, base = 3),
                     grade = contr.treatment(3, base = 2),
                     trt = contr.treatment(2, base = 2))
  ) %>%
    tidy_and_attach() %>%
    tidy_add_reference_rows()
}

#> # A tibble: 12 x 10
#>    term  variable var_class var_type contrasts reference_row estimate std.error
#>    <chr> <chr>    <chr>     <chr>    <chr>     <lgl>            <dbl>     <dbl>
#>  1 (Int… NA       NA        interce… NA        NA              -0.196     0.487
#>  2 stag… stage    factor    categor… contr.tr… FALSE           -0.138     0.455
#>  3 stag… stage    factor    categor… contr.tr… FALSE           -0.698     0.471
#>  4 stag… stage    factor    categor… contr.tr… TRUE            NA        NA    
#>  5 stag… stage    factor    categor… contr.tr… FALSE           -0.447     0.466
#>  6 grad… grade    factor    categor… contr.tr… FALSE            0.142     0.516
#>  7 grad… grade    factor    categor… contr.tr… TRUE            NA        NA    
#>  8 grad… grade    factor    categor… contr.tr… FALSE           -0.615     0.550
#>  9 trt1  trt      character categor… contr.SAS FALSE           -0.710     0.575
#> 10 trt2  trt      character categor… contr.SAS TRUE            NA        NA    
#> 11 grad… grade:t… NA        interac… NA        NA              -0.180     0.794
#> 12 grad… grade:t… NA        interac… NA        NA               1.63      0.803
#> # … with 2 more variables: statistic <dbl>, p.value <dbl>