Add term labels

Will add term labels in a label column, based on:

labels provided in labels argument if provided;

factor levels for categorical variables coded with treatment, SAS or sum contrasts;

variable labels when there is only one term per variable;

term name otherwise.

tidy_add_term_labels(
  x,
  labels = NULL,
  interaction_sep = " * ",
  model = tidy_get_model(x)
)

Arguments

x
a tidy tibble

labels
an optional named list or named vector of custom term labels

interaction_sep
separator for interaction terms

model
the corresponding model, if not attached to x

Details

If the variable_label column is not yet available in x, tidy_add_variable_labels() will be automatically applied. If the contrasts column is not yet available in x, tidy_add_contrasts() will be automatically applied.

It is possible to pass a custom label for any term in labels, including interaction terms.

See also

Examples

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes"))) %>%
  labelled::set_variable_labels(
    Class = "Passenger's class",
    Sex = "Sex"
  )

df %>%
  glm(Survived ~ Class * Age * Sex, data = ., weights = .$n, family = binomial) %>%
  tidy_and_attach() %>%
  tidy_add_term_labels()

#> # A tibble: 16 x 11
#>    term  variable var_class var_type estimate std.error statistic   p.value
#>    <chr> <chr>    <chr>     <chr>       <dbl>     <dbl>     <dbl>     <dbl>
#>  1 (Int~ NA       NA        interce~   3.56       0.507  7.01      2.36e-12
#>  2 Clas~ Class    character categor~  -1.74       0.589 -2.95      3.15e- 3
#>  3 Clas~ Class    character categor~  -3.71       0.531 -7.00      2.59e-12
#>  4 Clas~ Class    character categor~  -1.66       0.800 -2.07      3.83e- 2
#>  5 AgeC~ Age      character categor~  11.0      883.     0.0125    9.90e- 1
#>  6 SexM~ Sex      character categor~  -4.28       0.532 -8.05      8.36e-16
#>  7 Clas~ Class:A~ NA        interac~   3.53    1066.     0.00331   9.97e- 1
#>  8 Clas~ Class:A~ NA        interac~ -11.0      883.    -0.0125    9.90e- 1
#>  9 Clas~ Class:A~ NA        interac~  NA         NA     NA        NA       
#> 10 Clas~ Class:S~ NA        interac~   0.0680     0.671  0.101     9.19e- 1
#> 11 Clas~ Class:S~ NA        interac~   2.80       0.569  4.92      8.52e- 7
#> 12 Clas~ Class:S~ NA        interac~   1.14       0.820  1.38      1.66e- 1
#> 13 AgeC~ Age:Sex  NA        interac~   5.25    1091.     0.00481   9.96e- 1
#> 14 Clas~ Class:A~ NA        interac~  -1.19    1383.    -0.000860  9.99e- 1
#> 15 Clas~ Class:A~ NA        interac~  -4.57    1091.    -0.00419   9.97e- 1
#> 16 Clas~ Class:A~ NA        interac~  NA         NA     NA        NA       
#> # ... with 3 more variables: var_label <chr>, contrasts <chr>, label <chr>