Marginal Effects Estimation

Use effects::allEffects() to estimate "marginal effects" and return a tibble tidied in a way that could be used by broom.helpersfunctions. See vignette("functions-supported-by-effects", package = "effects") for a list of supported models.

tidy_all_effects(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x

a model

conf.int

logical indicating whether or not to include a confidence interval in the tidied output

conf.level

the confidence level to use for the confidence interval

...

additional parameters passed to effects::allEffects()

See also

Examples

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
mod <- glm(
  Survived ~ Class + Age + Sex,
  data = df, weights = df$n, family = binomial
)
tidy_all_effects(mod)
#>   variable   term  estimate  std.error  conf.low conf.high
#> 1    Class    1st 0.7965028 0.02932368 0.7330188 0.8480211
#> 2    Class    2nd 0.5857646 0.04297885 0.4998303 0.6667775
#> 3    Class    3rd 0.3981490 0.03198514 0.3374233 0.4621800
#> 4    Class   Crew 0.6240800 0.03847047 0.5462441 0.6959949
#> 5      Age  Adult 0.4801165 0.01668751 0.4475410 0.5128619
#> 6      Age  Child 0.7274982 0.04762469 0.6250690 0.8104309
#> 7      Sex Female 0.8404024 0.02178974 0.7929525 0.8786423
#> 8      Sex   Male 0.3188992 0.02736413 0.2678105 0.3747458
tidy_plus_plus(mod, tidy_fun = tidy_all_effects)
#> # A tibble: 8 × 16
#>   term   variable var_la…¹ var_c…² var_t…³ var_n…⁴ contr…⁵ contr…⁶ refer…⁷ label
#>   <chr>  <chr>    <chr>    <chr>   <chr>     <int> <chr>   <chr>   <lgl>   <chr>
#> 1 1st    Class    Class    charac… catego…       4 contr.… treatm… NA      1st  
#> 2 2nd    Class    Class    charac… catego…       4 contr.… treatm… NA      2nd  
#> 3 3rd    Class    Class    charac… catego…       4 contr.… treatm… NA      3rd  
#> 4 Crew   Class    Class    charac… catego…       4 contr.… treatm… NA      Crew 
#> 5 Adult  Age      Age      charac… dichot…       2 contr.… treatm… NA      Adult
#> 6 Child  Age      Age      charac… dichot…       2 contr.… treatm… NA      Child
#> 7 Female Sex      Sex      charac… dichot…       2 contr.… treatm… NA      Fema…
#> 8 Male   Sex      Sex      charac… dichot…       2 contr.… treatm… NA      Male 
#> # … with 6 more variables: n_obs <dbl>, n_event <dbl>, estimate <dbl>,
#> #   std.error <dbl>, conf.low <dbl>, conf.high <dbl>, and abbreviated variable
#> #   names ¹​var_label, ²​var_class, ³​var_type, ⁴​var_nlevels, ⁵​contrasts,
#> #   ⁶​contrasts_type, ⁷​reference_row