tidy_attach_model {broom.helpers}	R Documentation

Attach a full model to the tibble of model terms

Description

To facilitate the use of broom helpers with pipe, it is recommended to attach the original model as an attribute to the tibble of model terms generated by broom::tidy().

Usage

tidy_attach_model(x, model, .attributes = NULL)

tidy_and_attach(
  model,
  tidy_fun = tidy_with_broom_or_parameters,
  exponentiate = FALSE,
  ...
)

tidy_get_model(x)

tidy_detach_model(x)

Arguments

x
a tibble of model terms

model
a model to be attached/tidied

.attributes
named list of additional attributes to be attached to x

tidy_fun
option to specify a custom tidier function

exponentiate
logical indicating whether or not to exponentiate the coefficient estimates. This is typical for logistic, Poisson and Cox models, but a bad idea if there is no log or logit link; defaults to FALSE

...
other arguments passed to tidy_fun()

Details

tidy_attach_model() attach the model to a tibble already generated while tidy_and_attach() will apply broom::tidy() and attach the model.

Use tidy_get_model() to get the model attached to the tibble and tidy_detach_model() to remove the attribute containing the model.

See Also

Other tidy_helpers: tidy_add_coefficients_type(), tidy_add_contrasts(), tidy_add_estimate_to_reference_rows(), tidy_add_header_rows(), tidy_add_n(), tidy_add_reference_rows(), tidy_add_term_labels(), tidy_add_variable_labels(), tidy_disambiguate_terms(), tidy_identify_variables(), tidy_plus_plus(), tidy_remove_intercept(), tidy_select_variables()

Examples

mod <- lm(Sepal.Length ~ Sepal.Width + Species, data = iris)
tt <- mod %>%
  tidy_and_attach(conf.int = TRUE)
tt
tidy_get_model(tt)

[Package broom.helpers version 1.3.0.9001 Index]