Conditional Effects Estimation

Use ggeffects::ggpredict() to estimate "conditional effects" and return a tibble tidied in a way that could be used by broom.helpersfunctions. See https://strengejacke.github.io/ggeffects/ for a list of supported models.

tidy_ggpredict(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x

a model

conf.int

logical indicating whether or not to include a confidence interval in the tidied output

conf.level

the confidence level to use for the confidence interval

...

additional parameters passed to ggeffects::ggpredict()

See also

Examples

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
mod <- glm(
  Survived ~ Class + Age + Sex,
  data = df, weights = df$n, family = binomial
)
tidy_ggpredict(mod)
#> # A tibble: 8 × 6
#>   term   estimate std.error conf.low conf.high variable
#>   <chr>     <dbl>     <dbl>    <dbl>     <dbl> <chr>   
#> 1 1st       0.885     0.168    0.847     0.915 Class   
#> 2 2nd       0.736     0.167    0.668     0.795 Class   
#> 3 3rd       0.566     0.129    0.503     0.627 Class   
#> 4 Crew      0.766     0.159    0.706     0.817 Class   
#> 5 Child     0.957     0.298    0.926     0.976 Age     
#> 6 Adult     0.885     0.168    0.847     0.915 Age     
#> 7 Male      0.407     0.136    0.345     0.473 Sex     
#> 8 Female    0.885     0.168    0.847     0.915 Sex     
tidy_plus_plus(mod, tidy_fun = tidy_ggpredict)
#> # A tibble: 8 × 16
#>   term   variable var_la…¹ var_c…² var_t…³ var_n…⁴ contr…⁵ contr…⁶ refer…⁷ label
#>   <chr>  <chr>    <chr>    <chr>   <chr>     <int> <chr>   <chr>   <lgl>   <chr>
#> 1 1st    Class    Class    charac… catego…       4 contr.… treatm… NA      1st  
#> 2 2nd    Class    Class    charac… catego…       4 contr.… treatm… NA      2nd  
#> 3 3rd    Class    Class    charac… catego…       4 contr.… treatm… NA      3rd  
#> 4 Crew   Class    Class    charac… catego…       4 contr.… treatm… NA      Crew 
#> 5 Child  Age      Age      charac… dichot…       2 contr.… treatm… NA      Child
#> 6 Adult  Age      Age      charac… dichot…       2 contr.… treatm… NA      Adult
#> 7 Male   Sex      Sex      charac… dichot…       2 contr.… treatm… NA      Male 
#> 8 Female Sex      Sex      charac… dichot…       2 contr.… treatm… NA      Fema…
#> # … with 6 more variables: n_obs <dbl>, n_event <dbl>, estimate <dbl>,
#> #   std.error <dbl>, conf.low <dbl>, conf.high <dbl>, and abbreviated variable
#> #   names ¹​var_label, ²​var_class, ³​var_type, ⁴​var_nlevels, ⁵​contrasts,
#> #   ⁶​contrasts_type, ⁷​reference_row