Average Marginal Effects Estimation

Use margins::margins() to estimate "average marginal effects" and return a tibble tidied in a way that could be used by broom.helpersfunctions. See margins::margins() for a list of supported models.

tidy_margins(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x

a model

conf.int

logical indicating whether or not to include a confidence interval in the tidied output

conf.level

the confidence level to use for the confidence interval

...

additional parameters passed to margins::margins()

Note

When applying margins::margins(), custom contrasts are ignored. Treatment contrasts (stats::contr.treatment()) are applied to all categorical variables. Interactions are also ignored.

See also

Examples

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
mod <- glm(
  Survived ~ Class + Age + Sex,
  data = df, weights = df$n, family = binomial
)
tidy_margins(mod)
#> Error in data[[variable]]: object of type 'closure' is not subsettable
tidy_plus_plus(mod, tidy_fun = tidy_margins)
#> ✖ There was an error calling `tidy_fun()`. Most likely, this is because the
#> function supplied in `tidy_fun=` was misspelled, does not exist, is not
#> compatible with your object, or was missing necessary arguments (e.g. `conf.level=` or `conf.int=`). See error message below.
#> Error: Error in data[[variable]]: object of type 'closure' is not subsettable