Tidy a zeroinfl or a hurdle model

A tidier for models generated with pscl::zeroinfl() or pscl::hurdle(). Term names will be updated to be consistent with generic models. The original term names are preserved in an "original_term" column.

Usage

tidy_zeroinfl(x, conf.int = TRUE, conf.level = 0.95, component = NULL, ...)

Arguments

x

a pscl::zeroinfl() or a pscl::hurdle() model

conf.int

logical indicating whether or not to include a confidence interval in the tidied output

conf.level

the confidence level to use for the confidence interval

component

NULL or one of "all", "conditional", "zi", or "zero_inflated"

...

additional parameters passed to parameters::model_parameters()

See also

Examples

if (FALSE) { # interactive()
if (.assert_package("pscl", boolean = TRUE)) {
  library(pscl)
  mod <- zeroinfl(
    art ~ fem + mar + phd,
    data = pscl::bioChemists
  )

  mod %>% tidy_zeroinfl(exponentiate = TRUE)
}
}