Variable selector

This function will soon be removed from broom.helpers. Please consider cards::process_selectors() as an alternative.

Usage

.select_to_varnames(
  select,
  data = NULL,
  var_info = NULL,
  arg_name = NULL,
  select_single = FALSE
)

Arguments

select

A single object selecting variables, e.g. c(age, stage), starts_with("age")

data

A data frame to select columns from. Default is NULL

var_info

A data frame of variable names and attributes. May also pass a character vector of variable names. Default is NULL

arg_name

Optional string indicating the source argument name. This helps in the error messaging. Default is NULL.

select_single

Logical indicating whether the result must be a single variable. Default is FALSE

Value

A character vector of variable names

Details

Function takes select()-like inputs and converts the selector to a character vector of variable names. Functions accepts tidyselect syntax, and additional selector functions defined within the package