model_compute_terms_contributions {broom.helpers}	R Documentation

Compute a matrix of terms contributions

Description

Used for model_get_n(). For each row and term, equal 1 if this row should be taken into account in the estimate of the number of observations, 0 otherwise.

Usage

model_compute_terms_contributions(model)

## Default S3 method:
model_compute_terms_contributions(model)

Arguments

model
a model object

Details

This function does not cover lavaan models (NULL is returned).

See Also

Other model_helpers: model_get_assign(), model_get_coefficients_type(), model_get_contrasts(), model_get_model_frame(), model_get_model_matrix(), model_get_model(), model_get_nlevels(), model_get_n(), model_get_offset(), model_get_response(), model_get_terms(), model_get_weights(), model_get_xlevels(), model_identify_variables(), model_list_contrasts(), model_list_terms_levels(), model_list_variables()

Examples

mod <- lm(Sepal.Length ~ Sepal.Width, iris)
mod %>% model_compute_terms_contributions()

mod <- lm(hp ~ mpg + factor(cyl) + disp:hp, mtcars)
mod %>% model_compute_terms_contributions()

mod <- glm(
  response ~ stage * grade + trt,
  gtsummary::trial,
  family = binomial,
  contrasts = list(
    stage = contr.sum,
    grade = contr.treatment(3, 2),
    trt = "contr.SAS"
  )
)
mod %>% model_compute_terms_contributions()

mod <- glm(
  response ~ stage * trt,
  gtsummary::trial,
  family = binomial,
  contrasts = list(stage = contr.poly)
)
mod %>% model_compute_terms_contributions()

mod <- glm(
  Survived ~ Class * Age + Sex,
  data = Titanic %>% as.data.frame(),
  weights = Freq, family = binomial
)
mod %>% model_compute_terms_contributions()

d <- dplyr::as_tibble(Titanic) %>%
  dplyr::group_by(Class, Sex, Age) %>%
  dplyr::summarise(
    n_survived = sum(n * (Survived == "Yes")),
    n_dead = sum(n * (Survived == "No"))
  )
mod <- glm(cbind(n_survived, n_dead) ~ Class * Age + Sex, data = d, family = binomial)
mod %>% model_compute_terms_contributions()

[Package broom.helpers version 1.3.0.9001 Index]