Get the model matrix of a model

The structure of the object returned by stats::model.matrix() could slightly differ for certain types of models. model_get_model_matrix() will always return an object with the same structure as stats::model.matrix.default().

Usage

model_get_model_matrix(model, ...)

# S3 method for default
model_get_model_matrix(model, ...)

# S3 method for multinom
model_get_model_matrix(model, ...)

# S3 method for clm
model_get_model_matrix(model, ...)

# S3 method for brmsfit
model_get_model_matrix(model, ...)

# S3 method for glmmTMB
model_get_model_matrix(model, ...)

# S3 method for plm
model_get_model_matrix(model, ...)

# S3 method for biglm
model_get_model_matrix(model, ...)

# S3 method for model_fit
model_get_model_matrix(model, ...)

# S3 method for fixest
model_get_model_matrix(model, ...)

# S3 method for LORgee
model_get_model_matrix(model, ...)

# S3 method for betareg
model_get_model_matrix(model, ...)

# S3 method for cch
model_get_model_matrix(model, ...)

# S3 method for cch
model_get_terms(model, ...)

Arguments

model

a model object

...

additional arguments passed to stats::model.matrix()

Details

For models fitted with glmmTMB::glmmTMB(), it will return a model matrix taking into account all components ("cond", "zi" and "disp"). For a more restricted model matrix, please refer to glmmTMB::model.matrix.glmmTMB().

For plm::plm() models, constant columns are not removed.

See also

Examples

lm(hp ~ mpg + factor(cyl), mtcars) %>%
  model_get_model_matrix() %>%
  head()
#>                   (Intercept)  mpg factor(cyl)6 factor(cyl)8
#> Mazda RX4                   1 21.0            1            0
#> Mazda RX4 Wag               1 21.0            1            0
#> Datsun 710                  1 22.8            0            0
#> Hornet 4 Drive              1 21.4            1            0
#> Hornet Sportabout           1 18.7            0            1
#> Valiant                     1 18.1            1            0