List contrasts used by a model

List contrasts used by a model

model_list_contrasts(model)

# S3 method for default
model_list_contrasts(model)

Arguments

model

a model object

Value

A tibble with three columns:

variable: variable name

contrasts: contrasts used

contrasts_type: type of contrasts ("treatment", "sum", "poly", "helmert", "sdiff, "other" or "no.contrast")

reference: for variables with treatment, SAS or sum contrasts, position of the reference level

Details

For models with no intercept, no contrasts will be applied to one of the categorical variable. In such case, one dummy term will be returned for each level of the categorical variable.

See also

Examples

glm(
  am ~ mpg + factor(cyl),
  data = mtcars,
  family = binomial,
  contrasts = list(`factor(cyl)` = contr.sum)
) %>%
  model_list_contrasts()
#> # A tibble: 1 × 4
#>   variable    contrasts reference contrasts_type
#>   <chr>       <chr>         <int> <chr>         
#> 1 factor(cyl) contr.sum         3 sum