Select helper functions

Set of functions to supplement the tidyselect set of functions for selecting columns of data frames (and other items as well).

all_continuous() selects continuous variables

all_categorical() selects categorical (including "dichotomous") variables

all_dichotomous() selects only type "dichotomous"

all_interaction() selects interaction terms from a regression model

all_intercepts() selects intercept terms from a regression model

all_contrasts() selects variables in regression model based on their type of contrast

all_ran_pars() and all_ran_vals() for random-effect parameters and values from a mixed model (see vignette("broom_mixed_intro", package = "broom.mixed"))

Usage

all_continuous()

all_dichotomous()

all_categorical(dichotomous = TRUE)

all_interaction()

all_ran_pars()

all_ran_vals()

all_intercepts()

all_contrasts(contrasts_type = NULL)

Arguments

dichotomous

Logical indicating whether to include dichotomous variables. Default is TRUE

contrasts_type

type of contrast to select. When NULL, all variables with a contrast will be selected. Default is NULL. Select among contrast types c("treatment", "sum", "poly", "helmert", "other")

Value

A character vector of column names selected

Examples

if (FALSE) { # interactive()
glm(response ~ age * trt + grade, gtsummary::trial, family = binomial) %>%
  tidy_plus_plus(exponentiate = TRUE, include = all_categorical())

glm(response ~ age + trt + grade + stage,
  gtsummary::trial,
  family = binomial,
  contrasts = list(trt = contr.SAS, grade = contr.sum, stage = contr.poly)
) %>%
  tidy_plus_plus(
    exponentiate = TRUE,
    include = all_contrasts(c("treatment", "sum"))
  )
}