Listing of Supported Models

Listing of Supported Models

supported_models

Format

A data frame with one row per supported model

model

Model

notes

Notes

Supported models

model	notes

biglm::bigglm()

biglmm::bigglm()

brms::brm()	broom.mixed package required

cmprsk::crr()	Limited support. It is recommended to use tidycmprsk::crr() instead.

fixest::feglm()	May fail with R <= 4.0.

fixest::femlm()	May fail with R <= 4.0.

fixest::feNmlm()	May fail with R <= 4.0.

fixest::feols()	May fail with R <= 4.0.

gam::gam()

geepack::geeglm()

glmmTMB::glmmTMB()	broom.mixed package required

lavaan::lavaan()	Limited support for categorical variables

lfe::felm()

lme4::glmer()	broom.mixed package required

lme4::glmer.nb()	broom.mixed package required

lme4::lmer()	broom.mixed package required

MASS::glm.nb()

MASS::polr()

mgcv::gam()	Use default tidier broom::tidy() for smooth terms only, or gtsummary::tidy_gam() to include parametric terms

mice::mira	Limited support. If mod is a mira object, use tidy_plus_plus(mod, tidy_fun = function(x, ...) mice::pool(x) %>% mice::tidy(...))

nnet::multinom()

ordinal::clm()	Limited support for models with nominal predictors.

ordinal::clmm()	Limited support for models with nominal predictors.

parsnip::model_fit	Supported as long as the type of model and the engine is supported.

plm::plm()

rstanarm::stan_glm()	broom.mixed package required

stats::aov()	Reference rows are not relevant for such models.

stats::glm()

stats::lm()

stats::nls()	Limited support

survey::svycoxph()

survey::svyglm()

survey::svyolr()

survival::clogit()

survival::coxph()

survival::survreg()

tidycmprsk::crr()

VGAM::vglm()	Limited support. It is recommended to use tidy_parameters() as tidy_fun.