Add variable labels

Will add variable labels in a var_label column, based on:

labels provided in labels argument if provided;

variable labels defined in the original data frame with the label attribute (cf. labelled::var_label());

variable name otherwise.

tidy_add_variable_labels(
  x,
  labels = NULL,
  interaction_sep = " * ",
  model = tidy_get_model(x),
  quiet = FALSE,
  strict = FALSE
)

Arguments

x
a tidy tibble

labels
an optional named list or named vector of custom variable labels

interaction_sep
separator for interaction terms

model
the corresponding model, if not attached to x

quiet
logical argument whether broom.helpers should not return a message when requested output cannot be generated. Default is FALSE

strict
logical argument whether broom.helpers should return an error when requested output cannot be generated. Default is FALSE

Details

If the variable column is not yet available in x, tidy_identify_variables() will be automatically applied.

It is possible to pass a custom label for an interaction term in labels (see examples).

See also

Examples

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes"))) %>%
  labelled::set_variable_labels(
    Class = "Passenger's class",
    Sex = "Sex"
  )

df %>%
  glm(Survived ~ Class * Age * Sex, data = ., weights = .$n, family = binomial) %>%
  tidy_and_attach() %>%
  tidy_add_variable_labels(
    labels = list(
      "(Intercept)" = "Custom intercept",
      Sex = "Gender",
      "Class:Age" = "Custom label"
    )
  )

#> # A tibble: 16 x 10
#>    term  variable var_label var_class var_type var_nlevels estimate std.error
#>    <chr> <chr>    <chr>     <chr>     <chr>          <int>    <dbl>     <dbl>
#>  1 (Int… (Interc… Custom i… NA        interce…          NA   3.56       0.507
#>  2 Clas… Class    Passenge… character categor…           4  -1.74       0.589
#>  3 Clas… Class    Passenge… character categor…           4  -3.71       0.531
#>  4 Clas… Class    Passenge… character categor…           4  -1.66       0.800
#>  5 AgeC… Age      Age       character dichoto…           2  11.0      883.   
#>  6 SexM… Sex      Gender    character dichoto…           2  -4.28       0.532
#>  7 Clas… Class:A… Custom l… NA        interac…          NA   3.53    1066.   
#>  8 Clas… Class:A… Custom l… NA        interac…          NA -11.0      883.   
#>  9 Clas… Class:A… Custom l… NA        interac…          NA  NA         NA    
#> 10 Clas… Class:S… Passenge… NA        interac…          NA   0.0680     0.671
#> 11 Clas… Class:S… Passenge… NA        interac…          NA   2.80       0.569
#> 12 Clas… Class:S… Passenge… NA        interac…          NA   1.14       0.820
#> 13 AgeC… Age:Sex  Age * Ge… NA        interac…          NA   5.25    1091.   
#> 14 Clas… Class:A… Passenge… NA        interac…          NA  -1.19    1383.   
#> 15 Clas… Class:A… Passenge… NA        interac…          NA  -4.57    1091.   
#> 16 Clas… Class:A… Passenge… NA        interac…          NA  NA         NA    
#> # … with 2 more variables: statistic <dbl>, p.value <dbl>