tidy_disambiguate_terms {broom.helpers}	R Documentation

Disambiguate terms

Description

For mixed models, the term column returned by broom.mixed may have duplicated values for random-effect parameters and random-effect values. In such case, the terms could be disambiguated be prefixing them with the value of the group column. tidy_disambiguate_terms() will not change any term if there is no group column in x. The original term value is kept in a new column original_term.

Usage

tidy_disambiguate_terms(x, sep = ".", model = tidy_get_model(x), quiet = FALSE)

Arguments

x
a tidy tibble

sep
character, separator added between group name and term

model
the corresponding model, if not attached to x

quiet
logical argument whether broom.helpers should not return a message when requested output cannot be generated. Default is FALSE

See Also

Other tidy_helpers: tidy_add_coefficients_type(), tidy_add_contrasts(), tidy_add_estimate_to_reference_rows(), tidy_add_header_rows(), tidy_add_n(), tidy_add_reference_rows(), tidy_add_term_labels(), tidy_add_variable_labels(), tidy_attach_model(), tidy_identify_variables(), tidy_plus_plus(), tidy_remove_intercept(), tidy_select_variables()

Examples

if (require(lme4) & require(broom.mixed) & require(gtsummary)) {
  mod <- lme4::lmer(marker ~ stage + (1|grade) + (death|response), trial)
  mod %>% tidy_and_attach() %>% tidy_disambiguate_terms()
}

[Package broom.helpers version 1.3.0.9001 Index]