Disambiguate terms

For mixed models, the term column returned by broom.mixed may have duplicated values for random-effect parameters and random-effect values. In such case, the terms could be disambiguated be prefixing them with the value of the group column. tidy_disambiguate_terms() will not change any term if there is no group column in x. The original term value is kept in a new column original_term.

tidy_disambiguate_terms(x, sep = ".", model = tidy_get_model(x), quiet = FALSE)

Arguments

x
a tidy tibble

sep
character, separator added between group name and term

model
the corresponding model, if not attached to x

quiet
logical argument whether broom.helpers should not return a message when requested output cannot be generated. Default is FALSE

See also

Examples

if (require(lme4) & require(broom.mixed) & require(gtsummary)) {
  mod <- lme4::lmer(marker ~ stage + (1|grade) + (death|response), trial)
  mod %>% tidy_and_attach() %>% tidy_disambiguate_terms()
}

#> Loading required package: lme4

#> Loading required package: Matrix

#> Loading required package: broom.mixed

#> Registered S3 method overwritten by 'broom.mixed':
#>   method      from 
#>   tidy.gamlss broom

#> Loading required package: gtsummary

#> 
#> Attaching package: ‘gtsummary’

#> The following object is masked from ‘package:broom.helpers’:
#> 
#>     all_continuous

#> boundary (singular) fit: see ?isSingular

#> # A tibble: 9 x 7
#>   term            original_term       effect group  estimate std.error statistic
#>   <chr>           <chr>               <chr>  <chr>     <dbl>     <dbl>     <dbl>
#> 1 (Intercept)     (Intercept)         fixed  NA       0.772      0.179     4.32 
#> 2 stageT2         stageT2             fixed  NA       0.350      0.170     2.06 
#> 3 stageT3         stageT3             fixed  NA       0.272      0.187     1.45 
#> 4 stageT4         stageT4             fixed  NA       0.131      0.173     0.757
#> 5 grade.sd__(Int… grade.sd__(Interce… ran_p… grade    0.153     NA        NA    
#> 6 response.sd__(… response.sd__(Inte… ran_p… respo…   0.121     NA        NA    
#> 7 response.cor__… response.cor__(Int… ran_p… respo…   1         NA        NA    
#> 8 response.sd__d… response.sd__death  ran_p… respo…   0.0468    NA        NA    
#> 9 Residual.sd__O… Residual.sd__Obser… ran_p… Resid…   0.841     NA        NA