Identify the variable corresponding to each model coefficient

tidy_identify_variables() will add to the tidy tibble three additional columns: variable, var_class, var_type and var_nlevels.

tidy_identify_variables(x, model = tidy_get_model(x), quiet = FALSE)

Arguments

x

a tidy tibble

model

the corresponding model, if not attached to x

quiet

logical argument whether broom.helpers should not return a message when requested output cannot be generated. Default is FALSE

Details

It will also identify interaction terms and intercept(s).

var_type could be:

"continuous",

"dichotomous" (categorical variable with 2 levels),

"categorical" (categorical variable with 3 levels or more),

"intercept"

"interaction"

"ran_pars (random-effect parameters for mixed models)

"ran_vals" (random-effect values for mixed models)

"unknown" in the rare cases where tidy_identify_variables() will fail to identify the list of variables

For dichotomous and categorical variables, var_nlevels corresponds to the number of original levels in the corresponding variables.

See also

Examples

Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes"))) %>%
  glm(Survived ~ Class + Age * Sex, data = ., weights = .$n, family = binomial) %>%
  tidy_and_attach() %>%
  tidy_identify_variables()
#> # A tibble: 7 × 11
#>   term  varia…¹ var_c…² var_t…³ var_n…⁴ estim…⁵ std.e…⁶ stati…⁷  p.value conf.…⁸
#>   <chr> <chr>   <chr>   <chr>     <int>   <dbl>   <dbl>   <dbl>    <dbl>   <dbl>
#> 1 (Int… (Inter… NA      interc…      NA   2.18    0.176  12.4   3.08e-35   1.84 
#> 2 Clas… Class   charac… catego…       4  -1.03    0.200  -5.17  2.29e- 7  -1.43 
#> 3 Clas… Class   charac… catego…       4  -1.81    0.176 -10.3   7.79e-25  -2.16 
#> 4 Clas… Class   charac… catego…       4  -0.803   0.160  -5.03  4.99e- 7  -1.12 
#> 5 AgeC… Age     charac… dichot…       2  -0.110   0.335  -0.328 7.43e- 1  -0.759
#> 6 SexM… Sex     charac… dichot…       2  -2.62    0.151 -17.3   2.80e-67  -2.92 
#> 7 AgeC… Age:Sex NA      intera…      NA   1.90    0.433   4.39  1.12e- 5   1.04 
#> # … with 1 more variable: conf.high <dbl>, and abbreviated variable names
#> #   ¹​variable, ²​var_class, ³​var_type, ⁴​var_nlevels, ⁵​estimate, ⁶​std.error,
#> #   ⁷​statistic, ⁸​conf.low

lm(
  Sepal.Length ~ poly(Sepal.Width, 2) + Species,
  data = iris,
  contrasts = list(Species = contr.sum)
) %>%
  tidy_and_attach(conf.int = TRUE) %>%
  tidy_identify_variables()
#> # A tibble: 5 × 11
#>   term         varia…¹ var_c…² var_t…³ var_n…⁴ estim…⁵ std.e…⁶ stati…⁷   p.value
#>   <chr>        <chr>   <chr>   <chr>     <int>   <dbl>   <dbl>   <dbl>     <dbl>
#> 1 (Intercept)  (Inter… NA      interc…      NA  5.84    0.0359 163.    4.32e-166
#> 2 poly(Sepal.… Sepal.… nmatri… contin…      NA  4.27    0.568    7.52  5.32e- 12
#> 3 poly(Sepal.… Sepal.… nmatri… contin…      NA -0.0720  0.447   -0.161 8.72e-  1
#> 4 Species1     Species factor  catego…       3 -1.13    0.0647 -17.5   1.22e- 37
#> 5 Species2     Species factor  catego…       3  0.324   0.0593   5.46  2.02e-  7
#> # … with 2 more variables: conf.low <dbl>, conf.high <dbl>, and abbreviated
#> #   variable names ¹​variable, ²​var_class, ³​var_type, ⁴​var_nlevels, ⁵​estimate,
#> #   ⁶​std.error, ⁷​statistic