tidy_parameters {broom.helpers}	R Documentation

Tidy a model with parameters package

Description

Use parameters::model_parameters() to tidy a model and apply parameters::standardize_names(style = "broom") to the output

Usage

tidy_parameters(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x
a model

conf.int
logical indicating whether or not to include a confidence interval in the tidied output

conf.level
the confidence level to use for the confidence interval

...
additional parameters passed to parameters::model_parameters()

See Also

Other custom_tieders: tidy_with_broom_or_parameters()

Examples

if (require(parameters)) {
  lm(Sepal.Length ~ Sepal.Width + Species, data = iris) %>%
    tidy_parameters()
}

[Package broom.helpers version 1.3.0.9001 Index]