Tidy a model with parameters package

Use parameters::model_parameters() to tidy a model and apply parameters::standardize_names(style = "broom") to the output

Usage

tidy_parameters(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x

a model

conf.int

logical indicating whether or not to include a confidence interval in the tidied output

conf.level

the confidence level to use for the confidence interval

...

additional parameters passed to parameters::model_parameters()

Note

For betareg::betareg(), the component column in the results is standardized with broom::tidy(), using "mean" and "precision" values.

See also

Examples

if (FALSE) { # interactive()
if (.assert_package("parameters", boolean = TRUE)) {
  lm(Sepal.Length ~ Sepal.Width + Species, data = iris) |>
    tidy_parameters()
}
}