Tidy a model and compute additional informations

This function will apply sequentially:

tidy_and_attach()

tidy_identify_variables()

tidy_add_contrasts()

tidy_add_reference_rows()

tidy_add_estimate_to_reference_rows()

tidy_add_variable_labels()

tidy_add_term_labels()

tidy_add_header_rows()

tidy_remove_intercept()

tidy_detach_model()

tidy_plus_plus(
  model,
  tidy_fun = broom::tidy,
  conf.int = TRUE,
  exponentiate = FALSE,
  variable_labels = NULL,
  term_labels = NULL,
  add_reference_rows = TRUE,
  add_estimate_to_reference_rows = TRUE,
  add_header_rows = FALSE,
  show_single_row = NULL,
  intercept = FALSE,
  keep_model = FALSE,
  ...
)

Arguments

model
a model to be attached/tidied

tidy_fun
option to specify a custom tidier function

conf.int
should confidence intervals be computed? (see broom::tidy())

exponentiate
logical indicating whether or not to exponentiate the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.

variable_labels
a named list or a named vector of custom variable labels

term_labels
a named list or a named vector of custom term labels

add_reference_rows
should reference rows be added?

add_estimate_to_reference_rows
should an estimate value be added to reference rows?

add_header_rows
should header rows be added?

show_single_row
a vector indicating the names of binary variables that should be displayed on a single row, when add_header_rows is TRUE

intercept
should the intercept(s) be included?

keep_model
should the model be kept as an attribute of the final result?

...
other arguments passed to tidy_fun()

See also

Examples

ex1 <- lm(Sepal.Length ~ Sepal.Width + Species, data = iris) %>%
  tidy_plus_plus()
ex1

#> # A tibble: 4 x 14
#>   term  variable var_class var_type estimate std.error statistic   p.value
#>   <chr> <chr>    <chr>     <chr>       <dbl>     <dbl>     <dbl>     <dbl>
#> 1 Sepa… Sepal.W… numeric   continu…    NA        0.106      7.56  4.19e-12
#> 2 Spec… Species  factor    categor…     0       NA         NA    NA       
#> 3 Spec… Species  factor    categor…     1.46     0.112     13.0   3.48e-26
#> 4 Spec… Species  factor    categor…     1.95     0.100     19.5   2.09e-42
#> # … with 6 more variables: conf.low <dbl>, conf.high <dbl>, contrasts <chr>,
#> #   reference_row <lgl>, var_label <chr>, label <chr>

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(
    Survived = factor(Survived, c("No", "Yes"))
  ) %>%
  labelled::set_variable_labels(
    Class = "Passenger's class",
    Sex = "Gender"
  )

ex2 <- glm(
  Survived ~ Class + Age * Sex,
  data = df, weights = df$n,
  family = binomial
) %>%
  tidy_plus_plus(exponentiate = TRUE)
ex2

#> # A tibble: 9 x 14
#>   term  variable var_class var_type estimate std.error statistic   p.value
#>   <chr> <chr>    <chr>     <chr>       <dbl>     <dbl>     <dbl>     <dbl>
#> 1 Clas… Class    character categor…   1         NA        NA     NA       
#> 2 Clas… Class    character categor…   0.356      0.200    -5.17   2.29e- 7
#> 3 Clas… Class    character categor…   0.164      0.176   -10.3    7.79e-25
#> 4 Clas… Class    character categor…   0.448      0.160    -5.03   4.99e- 7
#> 5 AgeA… Age      character categor…   1         NA        NA     NA       
#> 6 AgeC… Age      character categor…   0.896      0.335    -0.328  7.43e- 1
#> 7 SexF… Sex      character categor…   1         NA        NA     NA       
#> 8 SexM… Sex      character categor…   0.0730     0.151   -17.3    2.80e-67
#> 9 AgeC… Age:Sex  NA        interac…  NA          0.433     4.39   1.12e- 5
#> # … with 6 more variables: conf.low <dbl>, conf.high <dbl>, contrasts <chr>,
#> #   reference_row <lgl>, var_label <chr>, label <chr>

if(requireNamespace("gtsummary")) {
  ex3 <- glm(
    response ~ poly(age, 3) + stage + grade * trt,
    na.omit(gtsummary::trial),
    family = binomial,
    contrasts = list(stage = contr.treatment(4, base = 3),
                     grade = contr.sum)
  ) %>%
    tidy_plus_plus(
      exponentiate = TRUE,
      variable_labels = c(age = "Age (in years)"),
      add_header_rows = TRUE,
      show_single_row = "trt",
      term_labels = c("poly(age, 3)3" = "Cubic age"),
      keep_model = TRUE
    )
  ex3
}

#> # A tibble: 17 x 15
#>    term  variable var_class var_type estimate std.error statistic p.value
#>    <chr> <chr>    <chr>     <chr>       <dbl>     <dbl>     <dbl>   <dbl>
#>  1 NA    age      nmatrix.3 continu…   NA        NA      NA        NA    
#>  2 poly… age      nmatrix.3 continu…   NA         2.33    1.29      0.196
#>  3 poly… age      nmatrix.3 continu…   NA         2.35    0.0894    0.929
#>  4 poly… age      nmatrix.3 continu…   NA         2.39   -0.295     0.768
#>  5 NA    stage    factor    categor…   NA        NA      NA        NA    
#>  6 stag… stage    factor    categor…    1.00      0.496   0.00963   0.992
#>  7 stag… stage    factor    categor…    0.487     0.502  -1.43      0.152
#>  8 stag… stage    factor    categor…    1        NA      NA        NA    
#>  9 stag… stage    factor    categor…    0.650     0.511  -0.843     0.399
#> 10 NA    grade    factor    categor…   NA        NA      NA        NA    
#> 11 grad… grade    factor    categor…    0.690     0.363  -1.02      0.307
#> 12 grad… grade    factor    categor…    0.796     0.383  -0.595     0.552
#> 13 grad… grade    factor    categor…    3.16     NA      NA        NA    
#> 14 trtD… trt      character categor…    1.25      0.344   0.648     0.517
#> 15 NA    grade:t… NA        interac…   NA        NA      NA        NA    
#> 16 grad… grade:t… NA        interac…   NA         0.484   1.55      0.120
#> 17 grad… grade:t… NA        interac…   NA         0.506   0.662     0.508
#> # … with 7 more variables: conf.low <dbl>, conf.high <dbl>, contrasts <chr>,
#> #   reference_row <lgl>, var_label <chr>, label <chr>, header_row <lgl>