Tidy a model and compute additional informations

This function will apply sequentially:

tidy_and_attach()

tidy_disambiguate_terms()

tidy_identify_variables()

tidy_add_contrasts()

tidy_add_reference_rows()

tidy_add_estimate_to_reference_rows()

tidy_add_variable_labels()

tidy_add_term_labels()

tidy_add_header_rows()

tidy_add_n()

tidy_remove_intercept()

tidy_select_variables()

tidy_add_coefficients_type()

tidy_detach_model()

tidy_plus_plus(
  model,
  tidy_fun = tidy_with_broom_or_parameters,
  conf.int = TRUE,
  exponentiate = FALSE,
  variable_labels = NULL,
  term_labels = NULL,
  interaction_sep = " * ",
  categorical_terms_pattern = "{level}",
  disambiguate_terms = TRUE,
  disambiguate_sep = ".",
  add_reference_rows = TRUE,
  no_reference_row = NULL,
  add_estimate_to_reference_rows = TRUE,
  add_header_rows = FALSE,
  show_single_row = NULL,
  add_n = TRUE,
  intercept = FALSE,
  include = everything(),
  keep_model = FALSE,
  quiet = FALSE,
  strict = FALSE,
  ...
)

Arguments

model

a model to be attached/tidied

tidy_fun

option to specify a custom tidier function

conf.int

should confidence intervals be computed? (see broom::tidy())

exponentiate

logical indicating whether or not to exponentiate the coefficient estimates. This is typical for logistic, Poisson and Cox models, but a bad idea if there is no log or logit link; defaults to FALSE.

variable_labels

a named list or a named vector of custom variable labels

term_labels

a named list or a named vector of custom term labels

interaction_sep

separator for interaction terms

categorical_terms_pattern

a glue pattern for labels of categorical terms with treatment or sum contrasts (see model_list_terms_levels())

disambiguate_terms

should terms be disambiguated with tidy_disambiguate_terms()? (default TRUE)

disambiguate_sep

separator for tidy_disambiguate_terms()

add_reference_rows

should reference rows be added?

no_reference_row

variables (accepts tidyselect notation) for those no reference row should be added, when add_reference_rows = TRUE

add_estimate_to_reference_rows

should an estimate value be added to reference rows?

add_header_rows

should header rows be added?

show_single_row

variables that should be displayed on a single row (accepts tidyselect notation), when add_header_rows is TRUE

add_n

should the number of observations be added?

intercept

should the intercept(s) be included?

include

variables to include. Accepts tidyselect syntax. Use - to remove a variable. Default is everything(). See also all_continuous(), all_categorical(), all_dichotomous() and all_interaction()

keep_model

should the model be kept as an attribute of the final result?

quiet

logical argument whether broom.helpers should not return a message when requested output cannot be generated. Default is FALSE

strict

logical argument whether broom.helpers should return an error when requested output cannot be generated. Default is FALSE

...

other arguments passed to tidy_fun()

See also

Examples

ex1 <- lm(Sepal.Length ~ Sepal.Width + Species, data = iris) %>%
  tidy_plus_plus()
ex1
#> # A tibble: 4 × 17
#>   term              variable  var_label var_class var_type var_nlevels contrasts
#>   <chr>             <chr>     <chr>     <chr>     <chr>          <int> <chr>    
#> 1 Sepal.Width       Sepal.Wi… Sepal.Wi… numeric   continu…          NA NA       
#> 2 Speciessetosa     Species   Species   factor    categor…           3 contr.tr…
#> 3 Speciesversicolor Species   Species   factor    categor…           3 contr.tr…
#> 4 Speciesvirginica  Species   Species   factor    categor…           3 contr.tr…
#> # … with 10 more variables: contrasts_type <chr>, reference_row <lgl>,
#> #   label <chr>, n_obs <dbl>, estimate <dbl>, std.error <dbl>, statistic <dbl>,
#> #   p.value <dbl>, conf.low <dbl>, conf.high <dbl>

df <- Titanic %>%
  dplyr::as_tibble() %>%
  dplyr::mutate(
    Survived = factor(Survived, c("No", "Yes"))
  ) %>%
  labelled::set_variable_labels(
    Class = "Passenger's class",
    Sex = "Gender"
  )

ex2 <- glm(
  Survived ~ Class + Age * Sex,
  data = df, weights = df$n,
  family = binomial
) %>%
  tidy_plus_plus(
    exponentiate = TRUE,
    add_reference_rows = FALSE,
    categorical_terms_pattern = "{level} / {reference_level}",
    add_n = TRUE
  )
ex2
#> # A tibble: 6 × 17
#>   term             variable var_label   var_class var_type var_nlevels contrasts
#>   <chr>            <chr>    <chr>       <chr>     <chr>          <int> <chr>    
#> 1 Class2nd         Class    Passenger'… character categor…           4 contr.tr…
#> 2 Class3rd         Class    Passenger'… character categor…           4 contr.tr…
#> 3 ClassCrew        Class    Passenger'… character categor…           4 contr.tr…
#> 4 AgeChild         Age      Age         character dichoto…           2 contr.tr…
#> 5 SexMale          Sex      Gender      character dichoto…           2 contr.tr…
#> 6 AgeChild:SexMale Age:Sex  Age * Gend… NA        interac…          NA NA       
#> # … with 10 more variables: contrasts_type <chr>, label <chr>, n_obs <dbl>,
#> #   n_event <dbl>, estimate <dbl>, std.error <dbl>, statistic <dbl>,
#> #   p.value <dbl>, conf.low <dbl>, conf.high <dbl>

ex3 <-
  glm(
    response ~ poly(age, 3) + stage + grade * trt,
    na.omit(gtsummary::trial),
    family = binomial,
    contrasts = list(
      stage = contr.treatment(4, base = 3),
      grade = contr.sum
    )
  ) %>%
  tidy_plus_plus(
    exponentiate = TRUE,
    variable_labels = c(age = "Age (in years)"),
    add_header_rows = TRUE,
    show_single_row = all_dichotomous(),
    term_labels = c("poly(age, 3)3" = "Cubic age"),
    keep_model = TRUE
  )
ex3
#> # A tibble: 17 × 19
#>    term   variable var_label var_class var_type var_nlevels header_row contrasts
#>    <chr>  <chr>    <chr>     <chr>     <chr>          <int> <lgl>      <chr>    
#>  1 NA     age      Age (in … nmatrix.3 continu…          NA TRUE       NA       
#>  2 poly(… age      Age (in … nmatrix.3 continu…          NA FALSE      NA       
#>  3 poly(… age      Age (in … nmatrix.3 continu…          NA FALSE      NA       
#>  4 poly(… age      Age (in … nmatrix.3 continu…          NA FALSE      NA       
#>  5 NA     stage    T Stage   factor    categor…           4 TRUE       contr.tr…
#>  6 stage1 stage    T Stage   factor    categor…           4 FALSE      contr.tr…
#>  7 stage2 stage    T Stage   factor    categor…           4 FALSE      contr.tr…
#>  8 stage3 stage    T Stage   factor    categor…           4 FALSE      contr.tr…
#>  9 stage4 stage    T Stage   factor    categor…           4 FALSE      contr.tr…
#> 10 NA     grade    Grade     factor    categor…           3 TRUE       contr.sum
#> 11 grade1 grade    Grade     factor    categor…           3 FALSE      contr.sum
#> 12 grade2 grade    Grade     factor    categor…           3 FALSE      contr.sum
#> 13 grade3 grade    Grade     factor    categor…           3 FALSE      contr.sum
#> 14 trtDr… trt      Chemothe… character dichoto…           2 NA         contr.tr…
#> 15 NA     grade:t… Grade * … NA        interac…          NA TRUE       NA       
#> 16 grade… grade:t… Grade * … NA        interac…          NA FALSE      NA       
#> 17 grade… grade:t… Grade * … NA        interac…          NA FALSE      NA       
#> # … with 11 more variables: contrasts_type <chr>, reference_row <lgl>,
#> #   label <chr>, n_obs <dbl>, n_event <dbl>, estimate <dbl>, std.error <dbl>,
#> #   statistic <dbl>, p.value <dbl>, conf.low <dbl>, conf.high <dbl>