tidy_with_broom_or_parameters {broom.helpers}	R Documentation

Tidy a model with broom or parameters

Description

Try to tidy a model with broom::tidy(). If it fails, will try to tidy the model using parameters::model_parameters() through tidy_parameters().

Usage

tidy_with_broom_or_parameters(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x
a model

conf.int
logical indicating whether or not to include a confidence interval in the tidied output

conf.level
the confidence level to use for the confidence interval

...
additional parameters passed to broom::tidy() or parameters::model_parameters()

See Also

Other custom_tieders: tidy_parameters()

[Package broom.helpers version 1.3.0.9001 Index]