Tidy a model with broom or parameters

Try to tidy a model with broom::tidy(). If it fails, will try to tidy the model using parameters::model_parameters() through tidy_parameters().

Usage

tidy_with_broom_or_parameters(x, conf.int = TRUE, conf.level = 0.95, ...)

Arguments

x

(a model object, e.g. glm)
A model to be tidied.

conf.int

(logical)
Whether or not to include a confidence interval in the tidied output.

conf.level

(numeric)
The confidence level to use for the confidence interval (between 0 ans 1).

...

Additional parameters passed to broom::tidy() or parameters::model_parameters().

See also